Release 0.5.0 (#35)
* doc: Improve some methods documentation

Improve documentation in app/Zone.php and app/Records.php.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* style: Change search form and table to improve visibility

Change search form to improve Search button position.
Change results table to improve visibility when use a lot of data.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* feat: Add Server Side processing to large DataTables

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* fix: Remove Server Side processing on Zone table

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* doc: Improve documentation of Zone model.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* refactor: Reordering Zone model (work 20160907)

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* fix: Fix Model Factories to add all fields on models

Before this modification Model factories only create some fields, now we
deal with all of them.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* feat: Complete properties rename on Zone model

Some properties has been renamed:
'updated' to 'has_modifications'
'master' to 'master_server'

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* doc: Models documentation

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* test: Some files couldn't be tested

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* feat: Add probind:import Artisan command

Add a class to parse a BIND file and get an array with the records.
Then probind:import call this class to create a Zone and Records.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* feat: Add user feedback to probind:import

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* feat: Add 'Import zone' feature to ProBIND

Import zone will get a HTTP uploaded file and call 'probind:import'
Artisan command to parse the file. Data will be imported to a Zone.

This is a first working version. Some improvements has included for
future TODO.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* refactor: Fix some Scrutinizer warnings

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* fix: Fix parser to deal with $ORIGIN tags

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* feat: Remove inherit domain from record's name

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* fix: Fix Zone creation on 'probind:import'

Before this commit Zone creation does not set neither 'custom_settings' nor
'serial'.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* test: Add FileDNSParser tests

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* test: Add test to parseToSeconds() function

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* refactor: Split parseToSeconds() into two functions

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* doc: Add 'Import Zone' feature to CHANGELOG

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* test: Add test to 'probind:import' console command

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* fix: Fix Zone editing

Zone name must not be edited. It's a parameter that affects to the rest
of records inside.

In example:

$ORIGIN domain.com.
ftp IN CNAME www.domain.com.

If we change the zone name:

$ORIGIN newdomain.com.
ftp IN CNAME www.domain.com.

It has no sense.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* refactor: Add a __construct() to FileDNSParser::class

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* fix: Fix scrutinizer issues on 'probind:import' command

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* feat: Add two functions to validate Zone names

One function will be used to normal zones and the other to reverse
zones.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* Scrutinizer Auto-Fixes

This commit consists of patches automatically generated for this project on https://scrutinizer-ci.com

* feat: Add a new Zone field to determine reverse zones

Change 'zones' table migration to add a new 'reverse_zone' field.
Add this feature to Seeders.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* feat: Add different Record Types for Reverse and Normal Zones

Some Records are not valid on Reverse or Normal zones. So this deal with
it.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* test: Add test to Reverse Zones code.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* test: Improve tests for FileDNSParser.

Add test for some lines that hasn't been tested before.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* fix: Change settings package to ARCANEV/Settings

Before this change we were using Torann/Registry as Settings Manager,
but this package only works for MySQL (See Issue #20).

I've changed to ARCANEV/Settings, I've changed all Setting calls, but
It's not tested, yet.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* fix: New Settings Manager change. Now use larapacks/setting.

There was some troubles with ARCANEDEV/settings and I've changed to
larapacks/setting package.

This is working great and could be tested.

ISSUE: There's an issue with cached values that are not changed when
call Setting::set().

Closes: #20

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* fix: Fix bug on larapacks/setting package

There was a bug on larapacks/setting package reported by me, See issue
20. I've updated larapacks/setting to version 1.0.5 which solves it.

Closes #20

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* feat: Create Users and Authentication Controller to ProBIND

Create User model and add authentication controllers / views to ProBIND

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* feat: Add LoginController to deal with login/logout

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* wip: Work at Office 20160929

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* test: Fix existing Http tests to use auth

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* feat: Add User management controller.

After adding User authentication it was necessary to add User's
management.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* test: Add tests for User management and authentication

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* doc: Modify CHANGELOG to add user authentication

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* refactor: Fix latest Scrutinizer issues.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* refactor: Fix some Scrutinizer issues.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* refactor: Fix UserController issue on Scrutinizer.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* refactor: Some fixes from Scrutinizer

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* Bump version 0.3.0 -> 0.4.0

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* feat: Create a 'create_package.sh' script

This script will create a release tar package.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* Remove `redis-server` from web container

Redis is served by a redis image, so it's not needed on the web
container.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* fix: Update Homestead to use latest version

Update [Homestead](https://laravel.com/docs/5.3/homestead) to latest `laravel/homestead` version.

Signed-off-by: Paco Orozco <paco@pacoorozco.info>

* Bump version 0.4.0 -> 0.5.0

Signed-off-by: Paco Orozco <paco@pacoorozco.info>